backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
alembic>=1.13.0
pydantic[email]>=2.10.0
pydantic-settings>=2.5.0
pyjwt>=2.8.0
bcrypt>=3.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
mysqlclient>=2.2.4
google-auth>=2.29.0
requests>=2.31.0
groq>=1.0.0
pdfplumber>=0.11.0
httpx>=0.27.0
